+++ student_class.rb
@@ -1,0 +1,15 @@
+require './personClass'
+
+class Student < Person
+  def initialize(age, classroom, name = 'Unknown', parent_permission: true)
+    super(age, name, parent_permission)
+    @classroom = classroom
+  end
+
+  def play_hooky
+    '¯(ツ)/¯'
+  end
+end
+
+student = Student.new(15, 'ibu', false, 'English 101')
+puts student.play_hooky
