+++ teacher_class.rb
@@ -1,0 +1,15 @@
+require './personClass'
+
+class Teacher < Person
+  def initialize(age, specialization, name = 'Unknown', parent_permission: true)
+    super(age, name, parent_permission)
+    @specialization = specialization
+  end
+
+  def can_use_services
+    true
+  end
+end
+
+teacher = Teacher.new(43, 'musa', 'physics')
+puts teacher.can_use_services
